+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.5.9 - Mar 11 2019
 - Add Gemfile support to Watchdog. When a `Gemfile` is added to a function, Watchdog will initialize it if the file is empty and run `bundle install` every time you modify it.
 - Watchdog will stream the output of commands instead of printing everything once it's done.
-
+- Disable caching for static files on Local
 
 ## 0.5.8 - Mar 11 2019
 - Bump dependency faastruby-rpc to v0.2.3